requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.9
faiss-cpu==1.7.4
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.229
langchainplus-sdk==0.0.20
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.21.12
pydantic==1.10.11
PyPDF2==3.0.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
SQLAlchemy==2.0.18
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
